Orden del código y generación de HTML y JS para filtros
